Rename "whiteboard_actions" package to "actions"
