+++ tests/fixtures/bad-config.yaml
@@ -4,6 +4,6 @@
   contact: foobar
   description: test config
   enabled: true
-  module: jbi.whiteboard_actions.default
+  module: jbi.actions.default
   parameters:
     jira_project_key: KEY
