+++ tests/unit/actions/test_default_with_assignee_and_status.py
@@ -1,5 +1,5 @@
 """
-Module for testing jbi/whiteboard_actions/extended.py functionality
+Module for testing jbi/actions/extended.py functionality
 """
 # pylint: disable=cannot-enumerate-pytest-fixtures
 from unittest import mock
@@ -8,8 +8,8 @@ from unittest.mock import MagicMock
 import pytest
 
 from jbi import Operation
+from jbi.actions import default_with_assignee_and_status as action
 from jbi.bugzilla import BugzillaBug, BugzillaWebhookRequest
-from jbi.whiteboard_actions import default_with_assignee_and_status as action
 
 
 def test_create_with_no_assignee(webhook_create_example, mocked_jira):
